Print header line every 50 comps
